srv_reader/requirements.txt
grpcio==1.43.0
grpcio-tools==1.43.0
Flask==2.2.2